fix: slight redesign of mutant selection logic
