+++ .gitignore
@@ -283,3 +283,4 @@ StrykerLogs/
 # End of https://www.gitignore.io/api/aspnetcore
 *.ncrunchsolution
 
+*.orig
