+++ src/Stryker.Core/Stryker.Core/Mutants/MutantOrchestrator.cs
@@ -32,6 +32,27 @@ namespace Stryker.Core.Mutants
         private IEnumerable<IMutator> _mutators { get; set; }
         private ILogger _logger { get; set; }
 
+        public static string MutantSelecterLogic
+        {
+            get { return helper; }
+        }
+
+        private const string helper = @"
+namespace Stryker
+{
+    public static class Environment
+    {
+        private static readonly int activeMutant;
+
+        static Environment()
+        {
+            activeMutant = int.Parse(System.Environment.GetEnvironmentVariable(""ActiveMutation"") ?? string.Empty);
+        }
+        public static int ID => activeMutant;
+    }
+}
+";
+
         /// <param name="mutators">The mutators that should be active during the mutation process</param>
         public MutantOrchestrator(IEnumerable<IMutator> mutators)
         {
